Fix the react dependency error when loading the inspector as custom element
