+++ .changeset/sweet-masks-raise.md
@@ -1,0 +1,5 @@
+---
+"jazz-tools": patch
+---
+
+Fix the react dependency error when loading the inspector as custom element
